requirements.txt
streamlit
pandas
plotly
python-dotenv
resend
qrcode
pillow